price_simulator/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.9.0            # via tensorboard, tensorflow
appdirs==1.4.3            # via virtualenv
aspy.yaml==1.3.0          # via pre-commit
astor==0.8.1              # via tensorflow
attrs==19.3.0
cfgv==3.0.0               # via pre-commit
cycler==0.10.0            # via matplotlib
filelock==3.0.12          # via virtualenv
gast==0.3.3               # via tensorflow
google-pasta==0.2.0       # via tensorflow
grpcio==1.30.0            # via tensorboard, tensorflow
h5py==2.10.0              # via keras, keras-applications
identify==1.4.11          # via pre-commit
importlib-metadata==1.5.0  # via markdown, pluggy, pre-commit, pytest, virtualenv
importlib-resources==1.0.2  # via pre-commit, virtualenv
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.2  # via tensorflow
keras==2.4.3
kiwisolver==1.1.0         # via matplotlib
markdown==3.2.2           # via tensorboard
matplotlib==3.1.3
more-itertools==8.2.0     # via pytest
nodeenv==1.3.5            # via pre-commit
numpy==1.18.1             # via h5py, keras, keras-applications, keras-preprocessing, matplotlib, pandas, scipy, tensorboard, tensorflow
networkx==2.5
packaging==20.1           # via pytest
pandas==1.0.1
pluggy==0.13.1            # via pytest
pre-commit==2.0.1
protobuf==3.12.2          # via tensorboard, tensorflow
py==1.8.1                 # via pytest
pyparsing==2.4.6          # via matplotlib, packaging
pytest==5.3.5
python-dateutil==2.8.1    # via matplotlib, pandas
pytz==2019.3              # via pandas
pyyaml==5.3               # via aspy.yaml, keras, pre-commit
scipy==1.4.1
six==1.14.0               # via absl-py, cycler, google-pasta, grpcio, h5py, keras-preprocessing, packaging, protobuf, python-dateutil, tensorboard, tensorflow, virtualenv
tabulate==0.8.7
tensorflow==2.2.0
termcolor==1.1.0          # via tensorflow
toml==0.10.0              # via pre-commit
virtualenv==20.0.1        # via pre-commit
wcwidth==0.1.8            # via pytest
werkzeug==1.0.1           # via tensorboard
wheel==0.34.2             # via tensorboard, tensorflow
wrapt==1.12.1             # via tensorflow
zipp==2.2.0               # via importlib-metadata
